requirements.txt
Django-tailwind
npm
mysqlclient
Pillow
